refactor(drawio): co-locate the mxfile format in @growi/remark-drawio
Adversarial review of #11522 flagged that the multi-page format was built
in apps/app (extractDrawioData) but detected in @growi/remark-drawio
(isMxfileData), with no shared source — the two halves could drift apart
silently.

Move both into a single packages/remark-drawio/src/utils/mxfile.ts so the
format is defined once. embed.ts imports isMxfileData from there; the barrel
re-exports only extractDrawioData (isMxfileData stays internal). apps/app now
imports extractDrawioData from @growi/remark-drawio. The end-to-end
save↔render round-trip test moves next to the shared module.

Refs #11522
