+++ packages/remark-drawio/src/index.ts
@@ -3,3 +3,5 @@ export * from './interfaces/graph-viewer.js';
 export * from './services/renderer/remark-drawio.js';
 export * from './utils/embed.js';
 export * from './utils/global.js';
+// Only the save-side builder is public; isMxfileData is internal to embed.ts.
+export { extractDrawioData } from './utils/mxfile.js';
